added POST and get cart

+++ ProductInfo.cs
@@ -5,6 +5,6 @@ namespace webreq
     public class ProductInfo
     {
         public long VariantId { get; set; }
-        public float Size { get; set; }
+        public double Size { get; set; }
     }
 }
